added new file scrollToTopProxy.js
a proxy that intercepts history.pushState calls and scrolls to the top of the page. Import the proxy in `main.js` to apply it globally. Use the proxy in the `handleClick` function of the Tours button to smoothly transition to the Tours page and scroll to the top of the page.